Add code for srf04. Should be able to deal with the JNI code.
Add also a dummy class to test without srf04.

+++ detection/src/main/java/esialrobotik/ia/detection/DetectionInterface.java
@@ -15,7 +15,7 @@ public interface DetectionInterface {
     //We never know right
     void stopDetection();
 
-    int[] ultraSoundDetection();
+    long[] ultraSoundDetection();
 
     List<PolarCoordinatePoint> getLidarDetection();
 }
